Remove redundant SignInCommandFactory and SignInEventFactory
